skille ut kalkulatoren i egen komponent fra part
