+++ src/components/_common/calculator/Calculator.less
