+++ src/components/_common/calculator/Result.less
