+++ src/components/_common/calculator/Result.tsx
