+++ src/components/parts/PartsMapper.tsx
@@ -23,7 +23,7 @@ import { NewsList } from './news-list/NewsList';
 import PublishingCalendar from './_legacy/publishing-calendar/PublishingCalendar';
 import { BEM, classNames } from '../../utils/classnames';
 import { HtmlArea } from './html-area/HtmlArea';
-import { Calculator } from './calculator/Calculator';
+import { CalculatorPart } from './calculator/Calculator';
 import { PageHeaderPart } from './page-header/PageHeaderPart';
 import { ButtonPart } from './button/ButtonPart';
 import { ProviderCardPart } from './provider-card/ProviderCardPart';
@@ -63,7 +63,7 @@ const partsWithOwnData: {
     [PartType.LinkList]: LinkList,
     [PartType.NewsList]: NewsList,
     [PartType.HtmlArea]: HtmlArea,
-    [PartType.Calculator]: Calculator,
+    [PartType.Calculator]: CalculatorPart,
     [PartType.PageHeader]: PageHeaderPart,
     [PartType.Button]: ButtonPart,
     [PartType.ProviderCard]: ProviderCardPart,
